# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdminJobResponse": {
        "properties": {
          "api_key_name": {
            "title": "Api Key Name",
            "type": "string"
          },
          "asset_refs": {
            "items": {
              "type": "string"
            },
            "title": "Asset Refs",
            "type": "array"
          },
          "billable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billable"
          },
          "candidate_types": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Types"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "deadline_at": {
            "title": "Deadline At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "jewelry_type_final": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jewelry Type Final"
          },
          "jewelry_type_requested": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jewelry Type Requested"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "matrix_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matrix Version"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "negative_prompt_snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt Snapshot"
          },
          "prompt_snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Snapshot"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "provider_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Job Id"
          },
          "reference_url_snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Url Snapshot"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submission_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Token"
          },
          "type_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Source"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "api_key_name",
          "service",
          "jewelry_type_requested",
          "jewelry_type_final",
          "type_source",
          "confidence",
          "candidate_types",
          "mock",
          "created_at",
          "updated_at",
          "completed_at",
          "deadline_at",
          "prompt_snapshot",
          "negative_prompt_snapshot",
          "reference_url_snapshot",
          "matrix_version",
          "provider",
          "submission_token",
          "provider_job_id",
          "billable",
          "asset_refs",
          "error_code",
          "error_message"
        ],
        "title": "AdminJobResponse",
        "type": "object"
      },
      "AssetRef": {
        "properties": {
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "index",
          "url",
          "mime"
        ],
        "title": "AssetRef",
        "type": "object"
      },
      "CandidateType": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "jewelry_type": {
            "$ref": "#/components/schemas/JewelryType"
          }
        },
        "required": [
          "jewelry_type",
          "confidence"
        ],
        "title": "CandidateType",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JewelryType": {
        "description": "Reconciled against the client's real matrix (Sheet1 header row) in Phase 0a\nStep 6. The client's sheet is the source of truth (docs/business-rules.md R9).",
        "enum": [
          "RING",
          "NECKLACE",
          "EARRING",
          "BRACELET",
          "BANGLE",
          "ANKLET",
          "HIPBELT"
        ],
        "title": "JewelryType",
        "type": "string"
      },
...
      "JobResponse": {
        "description": "Mirrors the GET /jobs/{job_id} example in docs/api-routes.md.\n\n`jewelry_type` maps to `Job.jewelry_type_final`, falling back to\n`Job.jewelry_type_requested` when the final type hasn't been set yet\n(e.g. while still `queued`/`classifying`) \u2014 this is a read-model\nconvenience so callers see the client-provided type immediately rather\nthan `null` until classification completes. The route that builds this\nfrom a `Job` (Step 4) is responsible for applying that fallback; this\nschema only declares the shape.",
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetRef"
            },
            "title": "Assets",
            "type": "array"
          },
          "candidate_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CandidateType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Types"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "deadline_at": {
            "title": "Deadline At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "jewelry_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JewelryType"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "service": {
            "$ref": "#/components/schemas/ServiceType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "type_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Source"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "service",
          "jewelry_type",
          "type_source",
          "confidence",
          "mock",
          "created_at",
          "updated_at",
          "completed_at",
          "deadline_at",
          "assets",
          "candidate_types",
          "error"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "classifying",
          "resolving",
          "submitting",
          "generating",
          "storing",
          "succeeded",
          "failed",
          "needs_input",
          "needs_review"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "ServiceType": {
        "description": "v1 values reconciled against the client's real matrix in Phase 0a Step 6: the\nsheet crosses 4 model/context categories with 2 styles (Traditional/Modern),\nnot the originally-assumed MODEL_SHOT/CATALOG_WHITE/LIFESTYLE/CLOSEUP_MACRO.",
        "enum": [
          "FEMALE_MODEL_TRADITIONAL",
          "FEMALE_MODEL_MODERN",
          "MALE_MODEL_TRADITIONAL",
          "MALE_MODEL_MODERN",
          "MANNEQUIN_TRADITIONAL",
          "MANNEQUIN_MODERN",
          "PRODUCT_STYLING_TRADITIONAL",
          "PRODUCT_STYLING_MODERN",
          "REMOVE_BG",
          "CHANGE_BG",
          "MIX_PIECES"
        ],
        "title": "ServiceType",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Jewellery Generation Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/jobs/{job_id}": {
      "get": {
        "operationId": "get_admin_job_api_v1_admin_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Admin Job"
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_route_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Route"
      }
    },
...
  }
}